DEV-1841 Disable redirect from unused endpoint
We should not be using the API endpoint that has been modified so it
should be safe to disable the redirect. Unfortunately it is probably
best tested _in situ_ in production.

One danger of leaving this redirect intact is that an attacker could
send a legitimate-looking link to an HMF employee, say requesting
support, then try to phish login information by spoofing one of our
actual login pages (nevermind that this application does not require
authentication itself).