DEV-1842 Do not display tooltips verbatim in HTML
The application allows a user to upload their own custom data and assign
an arbitrary name to it. The name is displayed in tooltips in the UI.
This allows injection of arbitrary code which could be used in an
attack.

This is best tested _in situ_ in production as there may be other
similar calls that have to be fixed, and I am not fully confident of the
fix without having been able to test in development.